restructured into scripts and viz, and merged scripts
