+++ viz/dempref_visuals.sh
